Удалил моковые данные и связанные с ними функции
