+++ js/api.js
@@ -1,5 +1,8 @@
+const GET_DATA_ADRESS = 'https://27.javascript.pages.academy/kekstagram/data';
+const SEND_DATA_ADRESS = 'https://27.javascript.pages.academy/kekstagram';
+
 const getData = (onSuccess, onFail) => {
-  fetch('https://27.javascript.pages.academy/kekstagram/data')
+  fetch(GET_DATA_ADRESS)
     .then((response) => {
       if(response.ok){
         return response.json();
@@ -18,7 +21,7 @@ const getData = (onSuccess, onFail) => {
 
 const sendData = (onSuccess, onFail, body) => {
   fetch(
-    'https://27.javascript.pages.academy/kekstagram',
+    SEND_DATA_ADRESS,
     {
       method: 'POST',
       body,
