+++ js/main.js
@@ -1,16 +1,9 @@
+import {showAlertMessage} from './util.js';
 import {renderBigPicture} from './big-picture.js';
 import {fillImgUploadForm, initUploadForm} from './img-upload.js';
 import {renderPhotos} from './render.js';
 import {getData} from './api.js';
 
-const errorElement = document.querySelector('#error').content.querySelector('section').cloneNode(true);
-
-const showAlertMessage = (message) =>{
-  errorElement.querySelector('.error__title').textContent = message;
-  const reloadButton = errorElement.querySelector('.error__button');
-  reloadButton.classList.add('hidden');
-  document.querySelector('body').appendChild(errorElement);
-};
 
 getData((photos) => {
   renderPhotos(photos);
