+++ js/util.js
@@ -1,22 +1,4 @@
-export const getRandomPositiveInteger = (a, b) => {
-  if (a < 0 || b < 0) {
-    return NaN;
-  }
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-
-export const getLengthCompare = (str, maxLength) => {
-  if(typeof maxLength !== 'number'){
-    return NaN;
-  }
-
-  return String(str).length <= maxLength;
-};
+const errorElement = document.querySelector('#error').content.querySelector('section').cloneNode(true);
 
 export const isEscapeKey = (evt) => evt.key === 'Escape';
 
@@ -34,4 +16,9 @@ export const validHashTag = (value) => {
   return true;
 };
 
-
+export const showAlertMessage = (message) =>{
+  errorElement.querySelector('.error__title').textContent = message;
+  const reloadButton = errorElement.querySelector('.error__button');
+  reloadButton.classList.add('hidden');
+  document.querySelector('body').appendChild(errorElement);
+};
